+++ package.json
@@ -4,8 +4,10 @@
   "description": "",
   "main": "index.js",
   "jest": {
+    "automock": false,
     "setupFiles": [
-      "jest-localstorage-mock"
+      "jest-localstorage-mock",
+      "./setupJest.js"
     ]
   },
   "scripts": {
@@ -20,12 +22,14 @@
     "@babel/core": "^7.12.17",
     "@babel/preset-env": "^7.12.17",
     "babel-loader": "^8.2.2",
+    "babel-polyfill": "^6.26.0",
     "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.2",
     "eslint": "^7.2.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
     "jest": "^26.6.3",
+    "jest-fetch-mock": "^3.0.3",
     "jest-localstorage-mock": "^2.4.7",
     "style-loader": "^2.0.0",
     "webpack": "^5.23.0",
@@ -36,6 +40,7 @@
     "bootstrap": "^4.6.0",
     "jquery": "^3.5.1",
     "phaser": "^3.52.0",
-    "popper.js": "^1.16.1"
+    "popper.js": "^1.16.1",
+    "regenerator-runtime": "^0.13.7"
   }
 }
