+++ setupJest.js
@@ -1,0 +1,1 @@
+global.fetch = require('jest-fetch-mock');
