+++ src/index.js
@@ -1,3 +1,4 @@
+import 'regenerator-runtime/runtime';
 import Game from './game';
 import './style.css';
 import 'bootstrap';
@@ -6,9 +7,9 @@ import domManip from './components/domManip';
 import addUser from './components/addUser';
 import checkUser from './components/checkUser';
 import getScores from './components/getScores';
+import sortArray from './components/sortArray';
 
 const dom = domManip();
-const id = 'ylaQJIceB2zpscPcvxIj';
 
 const {
   newUserContainer,
@@ -25,6 +26,7 @@ const {
   menu,
   exit,
   playersList,
+  players,
   back,
 } = dom;
 
@@ -84,7 +86,29 @@ leaderBoard.addEventListener('click', (e) => {
   buttonActions.classList.add('d-none');
   loading.classList.remove('d-none');
   playersList.classList.remove('d-none');
-  getScores(id);
+  getScores()
+  .then(function(array) {
+
+    loading.classList.add('d-none');
+    back.classList.remove('d-none');
+    while (players.firstChild) {
+      players.removeChild(players.lastChild);
+    }
+    const newArray = sortArray(array);
+    for (let i = 0; i < newArray.length; i++) {
+      const player = document.createElement('li');
+      player.classList.add('list-group-item');
+      const user = document.createElement('strong');
+      const score = document.createElement('span');
+      user.innerHTML = newArray[i].user + ': ';
+      score.innerHTML = newArray[i].score;
+      player.appendChild(user);
+      player.appendChild(score);
+      players.appendChild(player);
+    }
+    
+  });
+  
 
   back.addEventListener('click', (e) => {
     e.preventDefault();
